requirements.txt
asgiref==3.6.0
black==23.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
Deprecated==1.2.13
Django==4.2.1
django-environ==0.4.5
django-oauth-toolkit==2.2.0
django-ratelimit==4.0.0
djangorestframework==3.14.0
idna==3.4
jwcrypto==1.4.2
mypy-extensions==1.0.0
mysqlclient==2.1.1
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.1
pycparser==2.21
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
tomli==2.0.1
urllib3==2.0.2
wrapt==1.15.0
ratelimit==2.2.1

